Load shader module from file
